Add new builds pipeline to build the docker image
Refactor the CI pipeline to enable the new building pipeline which depends on the tests pipeline to succeed.
The new pipeline will build the go package and build the docker image. The docker image only get build if the active branch is the master branch and tag it with latest. It also tags the image if the pipeline got triggered by a new tag -> the tag has to be in the following format -> v1.0.0 (it will omit the 'v') to be used as an image tag.